@@ -1,30 +1,30 @@
 {
 
   "name": "split-frames",
-  "version": "2.4.5",
+  "version": "2.5.0",
   "description": "Split Buffer frames from streams",
 
   "type": "commonjs",
-  "main": "lib/main.js",
-  "typings": "lib/index.d.ts",
+  "typings": "./lib/cjs/main.d.cts",
+  "main": "./lib/cjs/main.cjs",
 
   "exports": {
     ".": {
       "require": {
-        "types": "./lib/index.d.ts",
-        "default": "./lib/main.js"
+        "types": "./lib/cjs/main.d.cts",
+        "default": "./lib/cjs/main.cjs"
       }
     }
   },
 
   "scripts": {
 
-    "build": "npx tsc --project \"./tsconfig.json\"",
+    "build": "node ./removeOldBuild.js && npx tsc --project \"./tsconfig.json\"",
 
-    "lint": "npx eslint ./lib/**/*.js ./test/**/*.js",
-    "check-requires": "npx used-deps-analyzer \"./package.json\" \"./lib\" --no-dev --overkill \"fs-extra\" \"node-promfs\"",
+    "lint": "npx eslint ./test/**/*.js",
+    "check-requires": "npx used-deps-analyzer \"./package.json\" \"./src\" --no-dev --overkill \"fs-extra\" \"node-promfs\"",
     "check-updates": "npx check-version-modules",
-    "unit-tests": "npx nyc --reporter=html --reporter=text mocha",
+    "unit-tests": "npm run build && npx nyc --reporter=html --reporter=text mocha",
 
     "tests": "npm run-script lint && npm run check-requires && npm run-script check-updates && npm run-script unit-tests",
     "ci": "npm run-script tests && npx nyc report --reporter=text-lcov | coveralls"
@@ -40,10 +40,10 @@
 
   "dependencies": {},
   "devDependencies": {
-    "@types/node": "20.5.0",
-    "check-version-modules": "1.4.1",
+    "@types/node": "20.5.7",
+    "check-version-modules": "1.5.0",
     "coveralls": "3.1.1",
-    "eslint": "8.47.0",
+    "eslint": "8.48.0",
     "husky": "8.0.3",
     "mocha": "10.2.0",
     "nyc": "15.1.0",
