@@ -1,6 +1,6 @@
-/// <reference path="../../lib/index.d.ts" />
+/// <reference path="../../lib/cjs/main.d.cts" />
 
-import Splitter = require("../../lib/main.js");
+import SplitFrames = require("../../lib/cjs/main.cjs");
 import { Readable } from "node:stream";
 
 const STX = 0x02, ETX = 0x03, DLE = 0x10, ACK = 0x06, NAK = 0x15, WAK = 0x13;
@@ -9,7 +9,7 @@ const stream = new Readable({
 	read () { }
 });
 
-stream.pipe(new Splitter({
+stream.pipe(new SplitFrames({
     "startWith": STX,
     "endWith": ETX,
     "specifics": {
