@@ -1,17 +1,22 @@
 "use strict";
 
-module.exports = (tags) => {
+// types & interfaces
 
-	let result = true;
+	// locals
+	import { tTag } from "../SplitFrames.js";
 
-		if (
-			"number" !== typeof tags && "object" !== typeof tags && !(tags instanceof Array || tags instanceof Buffer)
-		) {
+// module
+
+export default (tags: tTag): boolean => {
+
+	let result: boolean = true;
+
+		if ("number" !== typeof tags && "object" !== typeof tags) {
 			result = false;
 		}
 		else if ("object" === typeof tags && tags instanceof Array) {
 
-			for (let i = 0; i < tags.length; ++i) {
+			for (let i: number = 0; i < tags.length; ++i) {
 
 				if ("number" !== typeof tags[i] && "object" !== typeof tags[i] && !(tags[i] instanceof Buffer)) {
 					result = false; break;
